Rename seed to testSeed
